refactor styled components structure
