+++ src/layout/src/components/App/styled/AppNavigationInnerWrapper.js
@@ -1,0 +1,16 @@
+/*
+ * @flow
+ */
+
+import styled from 'styled-components';
+
+import AppContentInnerWrapper from './AppContentInnerWrapper';
+import { APP_CONTENT_PADDING } from '../../../../../style/Sizes';
+
+const AppNavigationInnerWrapper = styled(AppContentInnerWrapper)`
+  align-items: center;
+  flex-direction: ${({ vertical }) => (vertical ? 'column' : 'row')};
+  padding: 0 ${APP_CONTENT_PADDING}px;
+`;
+
+export default AppNavigationInnerWrapper;
